+++ addons/yat/src/enums/EDebugScreenItemPosition.cs
@@ -1,0 +1,9 @@
+namespace YAT.Enums;
+
+public enum EDebugScreenItemPosition
+{
+	TopLeft,
+	TopRight,
+	BottomLeft,
+	BottomRight
+}