+++ addons/yat/src/interfaces/IDebugScreenItem.cs
@@ -1,0 +1,8 @@
+namespace YAT.Interfaces;
+
+public interface IDebugScreenItem
+{
+	public string Title { get; set; }
+
+	public void Update();
+}