+++ addons/yat/src/scenes/debug_screen/items/cpu_info_item/CpuInfoItem.cs
@@ -3,23 +3,22 @@ using YAT.Interfaces;
 
 namespace YAT.Scenes;
 
-public partial class CpuInfo : PanelContainer, IMonitorComponent
+public partial class CpuInfoItem : PanelContainer, IDebugScreenItem
 {
-	public bool UseColors { get; set; }
+	public string Title { get; set; } = "CPU";
+
+	private Label _label;
+	private string _cpuInfo;
 
 	private readonly string _arch = Engine.GetArchitectureName();
 	private readonly string _cpuName = OS.GetProcessorName();
 	private readonly int _cpuCount = OS.GetProcessorCount();
-	private string _cpuInfo;
-
-	private Label _label;
 
 	public override void _Ready()
 	{
-		_label = GetNode<Label>("Label");
-
 		_cpuInfo = $"{_cpuName} {_arch} ({_cpuCount} cores)";
 
+		_label = GetNode<Label>("Label");
 		_label.Text = _cpuInfo;
 	}