+++ addons/yat/src/scenes/debug_screen/items/os_info_item/OsInfoItem.cs
@@ -3,26 +3,26 @@ using YAT.Interfaces;
 
 namespace YAT.Scenes;
 
-public partial class Os : PanelContainer, IMonitorComponent
+public partial class OsInfoItem : PanelContainer, IDebugScreenItem
 {
-	public bool UseColors { get; set; }
-
-	private readonly string _osName = OS.GetName();
-	private readonly string _osVersion = OS.GetVersion();
-	private readonly string _displayServerName = DisplayServer.GetName();
-	private readonly string _locale = OS.GetLocale();
-	private string _osInfo;
+	public string Title { get; set; } = "OS";
 
 	private Label _label;
+	private string _osInfo;
+
+	private readonly string _osName = OS.GetName();
+	private readonly string _locale = OS.GetLocale();
+	private readonly string _osVersion = OS.GetVersion();
+	private readonly string _displayServerName = DisplayServer.GetName();
 
 	public override void _Ready()
 	{
-		_label = GetNode<Label>("Label");
-
-		_osInfo = $"{_osName} {_osVersion} {_locale} {(_osName == _displayServerName
+		_osInfo = $"{_osName} {_osVersion} {_locale} {(
+			_osName == _displayServerName
 			? string.Empty
 			: $"({_displayServerName})")}";
 
+		_label = GetNode<Label>("Label");
 		_label.Text = _osInfo;
 	}