+++ addons/yat/src/scenes/debug_screen/items/scene_objects_info/SceneObjectsInfo.cs
@@ -3,9 +3,9 @@ using YAT.Interfaces;
 
 namespace YAT.Scenes;
 
-public partial class SceneObjects : PanelContainer, IMonitorComponent
+public partial class SceneObjectsInfo : PanelContainer, IDebugScreenItem
 {
-	public bool UseColors { get; set; }
+	public string Title { get; set; } = "SceneObjects";
 
 	private Label _label;
 
@@ -17,10 +17,15 @@ public partial class SceneObjects : PanelContainer, IMonitorComponent
 	public void Update()
 	{
 		var objects = Performance.GetMonitor(Performance.Monitor.ObjectCount);
-		var resources = Performance.GetMonitor(Performance.Monitor.ObjectResourceCount);
 		var nodes = Performance.GetMonitor(Performance.Monitor.ObjectNodeCount);
 		var orphans = Performance.GetMonitor(Performance.Monitor.ObjectOrphanNodeCount);
+		var resources = Performance.GetMonitor(Performance.Monitor.ObjectResourceCount);
 
-		_label.Text = $"Objects: {objects}\nResources: {resources}\nNodes: {nodes}\nOrphans: {orphans}";
+		_label.Text = string.Format("Objects: {0}\nResources: {1}\nNodes: {2}\nOrphans: {3}",
+			objects,
+			resources,
+			nodes,
+			orphans
+		);
 	}
 }